cmd/speaker: Separate speaker makefile from av/init
The makefile in init is mainly used for rv. Since we don’t have tests to check that the Makefile
still works for all types of netsender, we should create a seperate makefile for each netsender. This simplifies things. To install a netsender, one can just navigate to the directory in cmd and do sudo make install without remembering specific rules.