rv: Update audio pipeline to support I2S microphone (#440)
device/alsa: Fix poolbuffer reads

Reads from the pool buffer now get the chunk from the next call, and read from this before closing the chunk. Previously we were calling next, ignoring the returned chunk, and not closing it. This was causing buffer overflows.

rv: Use no delay for lexing

Using a delay of record period was incorrect behaviour, and this was causing issues between reading and writing rates, resulting in pool buffer overflows.

revid: set audio clipdur to 10 seconds

Hydrophone: Use a ticker to read per second (#440)

The input routine now uses a ticker to read in the equivalent of real time to ensure a more consitent flow of data. This ticker is also on a timeout to prevent blocking for mode changes.

Update comments and remove unused code.

This change cleans up some old comments and code that was incorrect or not being used.

Separate PCM and ADPCM Cases.

PCM and ADPCM cases were mistakenly merged, and have now been separated again to mantain functionality.