container/flv: add audio support (#45)
* container/flv: add audio support
We currently just add dummy audio, which does not allow us to have
anything other than silence.

* Fix formatting, add stereo switch

* av/container: split flv write into write audio and write video
This is so that audio and video can be written separately. Some adapters
are also provided to interface with io.Writer. A dummy audio decorator
is provided for when you don't want any audio.

* Add comments to consts

* Add audio to RTMP streaming test
Added audio functionality to the stream from file test for RTSP so that
we know it actually works. Also added a ADTS lexer, which may come in
handy.

* Fix formatting

* container/flv: Rename WriteVideo to Write
This is so that the previous interface is retained. Dummy audio is now
written by default again and can be enabled or disabled using
SetDummyAudio.

* Move variables in write into a var block

* Disable dummy audio in test